Drop infinite-precision soundness guarantees for verify, solve, and solve+.
These three queries no longer guarantee that their output will be sound with
respect to infinite-precision semantics when current-bitwidth is set to a
positive integer. The returned solutions are only guaranteed to be sound
with respect to the finite-precision semantics, as for the synthesize and
debug queries. Providing this guarantee was inefficient, and it is also no
longer necessary now that the client code is explicitly opting into the
finite precision semantics (with current-bitwidth being #f by default).