Closes #115.
The query in this issue causes Z3 to return algebraic numbers.
To avoid having to support these, we instead support decoding
Z3's approximate decimal representation of real numbers, and
when algebraic numbers are returned, we raise an error that
suggests re-solving the constraints using an instance of Z3
configured to print approximate decimal representation of (all)
real numbers.