Fix #147 (improve struct performance) and lift define-struct
This commit syncs changes from
https://github.com/racket/racket/commit/0de88f203d53c159dc968f65ee1c3d33e94ece3c
which speeds up struct performance when there are a lot of fields.
Note that there are other changes in `define-struct.rkt` that are left unsynced
which are probably safe to sync but have no real practical value in Rosette
(like supporting the #:authentic flag), so I leave them out. Following script

```
(define (generate n)
  (with-syntax ([(xs ...) (for/list ([i n]) (gensym))])
    (match-define-values (_ cpu _ _)
      (time-apply eval (list #'(struct foo (xs ...) #:transparent))))
    cpu))

;; warm up eval by padding with a dummy entry
(rest (for/list ([n (in-sequences '(0) (in-range 0 1001 100))])
        (list n (generate n))))
```

generates

```
'((0 3)
  (100 134)
  (200 474)
  (300 1122)
  (400 2156)
  (500 3378)
  (600 4857)
  (700 6840)
  (800 9492)
  (900 14984)
  (1000 22540))
```

before this commit. After applying the fix, the result is now:

```
'((0 40)
  (100 12)
  (200 68)
  (300 59)
  (400 69)
  (500 91)
  (600 116)
  (700 132)
  (800 224)
  (900 178)
  (1000 216))
```

This commit also fixes `define-struct` which was incorrectly exported, causing
it to be unlifted. In particular:

```
(define-struct a (b) #:transparent)

(define-symbolic b boolean?)

(a-b (if b (make-a 1) (make-a 2)))
```

in `#lang rosette/safe` used to fail, and this commit fixes the bug.