Add a library of useful bitvector operators.
This commit closes #152 and #136 by providing a set of new bitvector
operators including rotations, min/max, add/sub 1, converting a
bitvector to a list of bits, and more. It also fixes a bug in the
core bitvector library that caused binary and n-ary operators to
not emit enough type assertions in some cases.