Generalize bv extract / concat rewrite
This patch generalizes the rewrite rule for
`(extract _ _ (concat _ _))` to handle all cases where the result
depends only on the left or right argument of the `concat`. The old set
of rewrite rules handled the case where the result was exactly one of
the arguments of the concat; e.g. if `x` and `y` are `(bitvector 32)`,
then optimizing:

    (extract 63 32 (concat x y)) -> x

We can generalize this and do a little bit better for the cases where
the result depends only on the left or right argument of the concat, but
does not literally simplify to the left or right argument. For example:

    (extract 63 33 (concat x y)) -> (extract 31 1 x)

Before this patch, this would not have been simplified at all.

Compared to the previous version, this version emits no more new terms.