Update the default solver to the latest release of Z3.
This commit changes the Rosette install script to use the latest release of Z3
as the default solver instead of the custom Z3 version used so far. It also
changes some of the slower SDSL tests to use Boolector instead of Z3, to
compensate for the performance differences between the old and new Z3 versions.

The install script now works as follows. If the bin directory contains a symlink
to a Z3 binary, that symlink is left in place, so running `raco pkg install` has
no effect on the Z3 binary.  That is, Rosette will continue to use the symlink-ed
version of Z3. Otherwise, the install script downloads the latest release of Z3,
and this release replaces the Z3 binary (if any) that is currently in the bin
directory.