Instrument define-values (#168)
* Instrument define-values

The current tracer fails when a macro expands to define-values
at the top level due to an erroneous assumption that define-values
could not result in an error, which is not true as demonstrated in
macro-define.rkt. This PR thus adds define-values to a list of forms
to instrument. Note that define-values is different from other forms
because it's not an expression, so it needs a special handling.
The change also requires us to manually keep track the inferred name.

* lib must be in a separate file to trigger the bug