Revert "Try a minimized encoding that loses context."
This reverts commit d0ffeffcde000dcd3cbe42ab425ebc7748ba0912.

The problem with this encoding is that it loses context in a
way that breaks nested queries. Once we drop the context (i.e.,
assumes and asserts issued so far), then we no longer have a
way to tell whether a given query can ever be reached. This
was also a problem in Rosette 3.0, because nested queries
didn't account for the PC.