Update the syntax of define-symbolic[*] forms.
The new syntax uses the #:length keyword to specify the number of
symbolic constants to be created. This replaces the old syntax,
which allowed multidimensional lists to be created and used
ambiguous syntax.

To update client code to the new define-symbolic[*] forms, every
instance of

(define-symbolic id type [ k ])

should be replaced with

(define-symbolic id type #:length k)

Note that multidimensional lists are no longer supported. It's
easy for any application that needs this functionality
to implement it on top of one-dimensional lists.