Use pretty-write instead of pretty-print in print-forms
Using an example from the Rosette documentation, compare the output of
pretty-print (with print-as-expression set to #t):

(list
 'define
 '(bvmul2_bitfast x)
 (list 'bvadd 'x (list 'bvxor 'x (bv #x00 8))))

With the output of pretty-write:

(define (bvmul2_bitfast x) (bvadd x (bvxor x (bv #x00 8))))

The latter is properly formatted as Racket code and matches what is
shown in the Rosette guide.

Partially reverts 8fabaa8a0a07 ("Cleanup.").